+++ NEWS.md
@@ -1,5 +1,8 @@
 # dplyr (development version)
 
+* `arrange()` now works correctly when `across()` calls are used as the 2nd
+  (or more) ordering expression (#6495).
+
 * Joins now reference the correct column in `y` when a type error is thrown
   while joining on two columns with different names (#6465).